Fix whole-library review findings across all subsystems and backends
Correctness, security, and robustness fixes from a full-library review,
verified against affected tests on CPU and all four GPU backends
(CUDA RTX 5070, ROCm/Vulkan on Radeon 890M, OneAPI).

Core / autograd / ops:
- CPU conv2d_backward_input: fix wrong grad_output packing (channel-major
  vs the position-major layout the GEMM reads) that corrupted input grads
  for C_out>1, spatial>1; add a non-uniform-grad conv2d gradcheck that
  actually exercises it (plain sum() masks the bug with a uniform cotangent).
- Autograd AsStridedBackward/QrBackward create_graph adjoints; engine no
  longer pollutes a non-leaf root's .grad; LogSumExp no-dim OOB guard;
  iterative (stack-safe) jvp traversal.
- ops: expand GPU-path shape validation, one_hot/bincount index bounds,
  reject negative repeat/tile, float_power(0,0)=1, histogram zero-width
  range, corrcoef 1-D scalar shape, masked_scatter broadcast; dispatch_table
  OpId bounds; Tensor dtype predicates; MemoryManager keyed on TensorImpl.

CPU backend:
- maxpool1d/3d padding validation; adaptive-pool floor->ceil window (incl.
  AVX512); SIMD exp two-step reconstruction (correct band/subnormal range);
  double-precision batchnorm Welford accumulators.

NN / models:
- loss rank checks, CTC mean reduction, attention causal+padding, MoE router
  on-graph gate, dropout p==1 graph, LPPool, quantization round-half-to-even.
- T5 additive/reshaped masks + RMSNorm + gated FFN + weight tying, GPT-2 tie,
  hub SafeTensors extension, ConvNeXt LayerScale, ALBERT/ELECTRA masks/tying,
  AlexNet widths, Swin features, Mask R-CNN NMS + mask-on-decoded-boxes +
  low-quality anchors + empty-detection handling, Faster R-CNN ROIAlign
  stride, U-Net decoder concat, YOLOv5 decode, MobileNetV3 SE gate.
- Mask R-CNN RPN never returns zero proposals when anchors exist (fixes a
  crash on untrained deep backbones whose exploding features drive all boxes
  off-image); ROIHead robust to empty proposals; ForwardShape tests assert
  the deterministic output contract instead of a flaky positive count.

JIT / lazy / lite:
- ExtendedFusionPass inserts the fused node; tracing records graph breaks for
  unmapped ops; topological sort after node-appending passes; fused softmax
  axis; contiguity/device checks; lazy merge_graphs canonical resolution +
  cycle guard; gated Float64-safe Div->Mul; MLIR where broadcast; memory
  planner overflow; serialization Device::Type bounds; lite input validation.

Distributed:
- gradient-compression deadlock, atomic checkpoint writes + read checks +
  overflow guard, RREF ownership, RPC auth on non-loopback, fd/connection
  leak + cap, elastic rendezvous/leave races, health-monitor timeout.

Backends (correctness / no CPU fallbacks):
- CUDA: in-place op contiguity+dtype guards, bicubic coeff, repeat_interleave
  dim, sync return checks, where/masked_scatter broadcast, squeeze dim,
  gather bounds, fused int64 grid/pointers, handle-pool shutdown, graph
  capture, device-call checks.
- ROCm: GRU sync (UAF), dropout RNG seed, Float64 accumulators, IEEE-754
  NaN/Inf under -ffinite-math, non-contiguous flat ops, flash-attn grid;
  scatter/scatter_add shape-contract validation.
- OneAPI: element-wise/masked/squeeze contiguity, FFT BWD_DISTANCE, native
  Float64 CTC + 1-D targets, Int<->Int cast, avgpool1d count_include_pad,
  embedding neg-index, free-after-wait, norm forward contiguity, linspace
  complex, grid_sample contiguity.
- Vulkan: scatter_add/reduce contiguity + shape-contract validation, gather
  inner-stride, CSR sorted columns, F32 embedding-backward atomic-float gate,
  reduction-tree subgroup handling.

+++ python/tenzor/data.py
@@ -795,7 +795,7 @@ def _default_collate_impl(batch: list, _current_depth: int) -> Any:
         return _core.stack(batch, 0)
 
     # Numeric scalars: stack into a 1-D tensor matching PyTorch's default
-    # collate (int/bool -> int64, float -> float32).  Without this, the common
+    # collate (int/bool -> int64, float -> float64).  Without this, the common
     # ``Dataset.__getitem__ -> (tensor, int_label)`` pattern left the labels as
     # a raw Python list, breaking downstream loss/metric calls.  `bool` is a
     # subclass of `int`, so the int branch covers it (cast to int64).
@@ -804,7 +804,7 @@ def _default_collate_impl(batch: list, _current_depth: int) -> Any:
         return _core.Tensor.from_numpy(_np.asarray(batch, dtype=_np.int64))
     if isinstance(elem, float):
         import numpy as _np
-        return _core.Tensor.from_numpy(_np.asarray(batch, dtype=_np.float32))
+        return _core.Tensor.from_numpy(_np.asarray(batch, dtype=_np.float64))
 
     # Fallback for strings / other Python objects.
     return batch
